requirements.txt
numpy
pillow 
Image  
